Moved and renamed some files to function alongside libesphttp as a submodule.
